Adds impementations for actual game assets
This implements all 18 droppables, including their sprites and
collision shapes and the raw GIMP assets used to create the sprites.

It is a large commit that includes a lot of asset files that probably
shouldn't be comitted to the repository directly, but the tight scope
of the project means I'm willing to just let the assets live here.